python/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
faiss-cpu>=1.8.0
chromadb>=0.5.0
numpy>=1.26.0
httpx>=0.27.0
python-multipart>=0.0.12
pytest>=8.3.0
